+++ Content.Server/Power/Components/ApcPowerReceiverComponent.cs
@@ -1,17 +1,6 @@
-using System;
-using System.Diagnostics.CodeAnalysis;
 using Content.Server.Power.NodeGroups;
 using Content.Server.Power.Pow3r;
-using Content.Shared.Examine;
 using Content.Shared.Power;
-using Robust.Shared.GameObjects;
-using Robust.Shared.IoC;
-using Robust.Shared.Localization;
-using Robust.Shared.Maths;
-using Robust.Shared.Physics;
-using Robust.Shared.Serialization.Manager.Attributes;
-using Robust.Shared.Utility;
-using Robust.Shared.ViewVariables;
 
 namespace Content.Server.Power.Components
 {
@@ -21,7 +10,7 @@ namespace Content.Server.Power.Components
     /// </summary>
     [RegisterComponent]
 #pragma warning disable 618
-    public sealed class ApcPowerReceiverComponent : Component, IExamine
+    public sealed class ApcPowerReceiverComponent : Component
 #pragma warning restore 618
     {
         [Dependency] private readonly IEntityManager _entMan = default!;
@@ -88,16 +77,6 @@ namespace Content.Server.Power.Components
             {
                 appearance.SetData(PowerDeviceVisuals.Powered, Powered);
             }
-        }
-
-        ///<summary>
-        ///Adds some markup to the examine text of whatever object is using this component to tell you if it's powered or not, even if it doesn't have an icon state to do this for you.
-        ///</summary>
-        public void Examine(FormattedMessage message, bool inDetailsRange)
-        {
-            message.AddMarkup(Loc.GetString("power-receiver-component-on-examine-main",
-                                            ("stateText", Loc.GetString( Powered ? "power-receiver-component-on-examine-powered" :
-                                                                                   "power-receiver-component-on-examine-unpowered"))));
         }
     }
