+++ Content.Server/Power/EntitySystems/BatterySystem.cs
@@ -1,6 +1,6 @@
 using Content.Server.Power.Components;
+using Content.Shared.Examine;
 using JetBrains.Annotations;
-using Robust.Shared.GameObjects;
 
 namespace Content.Server.Power.EntitySystems
 {
@@ -11,8 +11,31 @@ namespace Content.Server.Power.EntitySystems
         {
             base.Initialize();
 
+            SubscribeLocalEvent<ExaminableBatteryComponent, ExaminedEvent>(OnExamine);
+
             SubscribeLocalEvent<NetworkBatteryPreSync>(PreSync);
             SubscribeLocalEvent<NetworkBatteryPostSync>(PostSync);
+        }
+
+        private void OnExamine(EntityUid uid, ExaminableBatteryComponent component, ExaminedEvent args)
+        {
+            if (!TryComp<BatteryComponent>(uid, out var batteryComponent))
+                return;
+            if (args.IsInDetailsRange)
+            {
+                var effectiveMax = batteryComponent.MaxCharge;
+                if (effectiveMax == 0)
+                    effectiveMax = 1;
+                var chargeFraction = batteryComponent.CurrentCharge / effectiveMax;
+                var chargePercentRounded = (int) (chargeFraction * 100);
+                args.PushMarkup(
+                    Loc.GetString(
+                        "examinable-battery-component-examine-detail",
+                        ("percent", chargePercentRounded),
+                        ("markupPercentColor", "green")
+                    )
+                );
+            }
         }
 
         private void PreSync(NetworkBatteryPreSync ev)
